requirements.txt
pysodium==0.7.5
argon2-cffi==20.1.0
